moved constants up the tree
